requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
numpy==2.1.3
oauthlib==3.2.2
openai==1.55.0
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.0
pydantic_core==2.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
